requirements.txt
annotated-types==0.7.0
anyio==3.7.1
apprise==1.5.0
APScheduler==3.10.4
Authlib==1.2.1
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
emoji==2.14.1
fastapi==0.103.2
feedparser==6.0.11
greenlet==3.2.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.1.5
Markdown==3.8
MarkupSafe==3.0.2
oauthlib==3.2.2
plombery==0.4.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.23.2
websockets==11.0.3
